requirements.txt
Flask==3.0.2
google-generativeai==0.3.2
fpdf==1.7.2
beautifulsoup4==4.12.3
requests==2.31.0
gunicorn==21.2.0 